+++ src/constants/index.ts
@@ -1,2 +1,2 @@
-export * from './tags'
 export * from './types'
+export * from './events'