+++ src/constants/tags.ts
@@ -1,4 +1,0 @@
-export const TAGS = {
-  HomeController: 'HomeController',
-  UserController: 'UserController'
-}