+++ src/constants/types.ts
@@ -1,5 +1,7 @@
-export const TYPES = {
+export const Types = {
   UserService: Symbol('UserService'),
+  GuildService: Symbol('GuildService'),
+  SocketService: Symbol('SocketService'),
   DatabaseProvider: Symbol('DatabaseProvider'),
   ConfigProvider: Symbol('ConfigProvider')
 }