+++ src/ioc/loader.ts
@@ -1,2 +1,5 @@
-import '../controller/user'
+import '../controllers/user'
 import '../services/user'
+import '../controllers/guild'
+import '../services/guild'
+import '../services/socket'